Former-commit-id: b14e2f874f0590c596641dbb57cb8f155f91a69d

+++ Comp/nuke/batchrender/batchrender.py
@@ -11,6 +11,7 @@ import logging
 import datetime
 import shutil
 import time
+import locale
 from subprocess import Popen, PIPE, call
 import multiprocessing
 
@@ -20,15 +21,16 @@ from PySide.QtGui import QMainWindow, QApplication, QFileDialog
 from ui_mainwindow import Ui_MainWindow
 
 
-__version__ = '0.5.0'
+__version__ = '0.6.4'
 EXE_PATH = os.path.join(os.path.dirname(__file__), 'batchrender.exe')
+OS_ENCODING = locale.getdefaultlocale()[1]
 
 reload(sys)
 sys.setdefaultencoding('UTF-8')
 
 
 class Config(dict):
-    """Config file as dict that automatic write and read json file."""
+    """A config file can be manipulated that automatic write and read json file on disk."""
 
     default = {
         'NUKE': r'C:\Program Files\Nuke10.0v4\Nuke10.0.exe',
@@ -76,11 +78,12 @@ class Config(dict):
 
 def change_dir(dir_):
     """Try change currunt working directory."""
+
     try:
-        os.chdir(dir_)
+        os.chdir(get_unicode(dir_))
     except WindowsError:
         print(sys.exc_info()[2])
-    print(u'工作目录改为: {}'.format(os.getcwd()).encode('UTF-8'))
+    print(u'工作目录改为: {}'.format(get_unicode(os.getcwd())))
 
 
 class SingleInstanceException(Exception):
@@ -150,6 +153,7 @@ class BatchRender(multiprocessing.Process):
     def set_logger(self):
         """Set logger for this process."""
 
+        self.rotate_log()
         self._logger = logging.getLogger(__name__)
         self._logger.setLevel(self.LOG_LEVEL)
         handler = logging.FileHandler(self.LOG_FILENAME)
@@ -157,7 +161,6 @@ class BatchRender(multiprocessing.Process):
             '[%(asctime)s]\t%(levelname)10s:\t%(message)s')
         handler.setFormatter(formatter)
         self._logger.addHandler(handler)
-        self.rotate_log()
 
     def continuous_render(self):
         """Loop batch rendering as files exists."""
@@ -168,21 +171,16 @@ class BatchRender(multiprocessing.Process):
     def rotate_log(self):
         """Rotate existed logfile if needed."""
 
-        if os.path.isfile(self.LOG_FILENAME):
-            if os.stat(self.LOG_FILENAME).st_size > 10000:
-                logname = os.path.splitext(self.LOG_FILENAME)[0]
-                # Remove oldest logfile.
-                if os.path.exists(u'{}.{}.log'.format(logname, 5)):
-                    os.remove(u'{}.{}.log'.format(logname, 5))
-                # Rename else.
-                for i in range(5)[:0:-1]:
-                    old_name = u'{}.{}.log'.format(logname, i)
-                    new_name = u'{}.{}.log'.format(logname, i + 1)
-                    if os.path.exists(old_name):
-                        os.rename(old_name, new_name)
-                if os.path.exists(self.LOG_FILENAME):
-                    os.rename(self.LOG_FILENAME,
-                              '{}.{}.log'.format(logname, 1))
+        prefix = os.path.splitext(self.LOG_FILENAME)[0]
+        if os.path.isfile(self.LOG_FILENAME) and os.stat(self.LOG_FILENAME).st_size > 10000:
+            for i in range(4, 0, -1):
+                old_name = u'{}.{}.log'.format(prefix, i)
+                new_name = u'{}.{}.log'.format(prefix, i + 1)
+                if os.path.exists(old_name):
+                    if os.path.exists(new_name):
+                        os.remove(new_name)
+                    os.rename(old_name, new_name)
+            os.rename(self.LOG_FILENAME, old_name)
 
     def batch_render(self):
         """Render all renderable file in dir."""
@@ -213,8 +211,8 @@ class BatchRender(multiprocessing.Process):
     def call_nuke(self, f):
         """Open a nuke subprocess for rendering file."""
 
-        _time = datetime.datetime.now()
-        _file = Files.lock(f)
+        current_time = datetime.datetime.now()
+        nk_file = Files.lock(f)
 
         _proxy = '-p ' if self._config['PROXY'] else '-f '
         _priority = '-c 8G --priority low ' if self._config['LOW_PRIORITY'] else ''
@@ -224,11 +222,11 @@ class BatchRender(multiprocessing.Process):
             _priority,
             _cont,
             NUKE=self._config['NUKE'],
-            f=_file,
+            f=nk_file
         )
         self._logger.debug(u'命令: %s', cmd)
         print(cmd)
-        _proc = Popen(cmd, stderr=PIPE)
+        _proc = unicode_popen(cmd, stderr=PIPE)
         self._queue.put(_proc.pid)
         _stderr = _proc.communicate()[1]
         _stderr = fanyi(_stderr)
@@ -245,7 +243,7 @@ class BatchRender(multiprocessing.Process):
         self._logger.info(
             u'%s: 结束渲染 耗时 %s %s',
             f,
-            timef((datetime.datetime.now() - _time).total_seconds()),
+            timef((datetime.datetime.now() - current_time).total_seconds()),
             u'退出码: {}'.format(_rtcode) if _rtcode else u'正常退出',
         )
 
@@ -259,14 +257,14 @@ class BatchRender(multiprocessing.Process):
                 self._logger.error(u'%s: 连续渲染错误超过3次,不再进行重试。', f)
             elif os.path.isfile(f):
                 # Retry, use new version.
-                os.remove(_file)
+                os.remove(nk_file)
             else:
                 # Retry, use this version.
-                os.rename(_file, f)
+                os.rename(nk_file, f)
         else:
             # Normal exit.
             if not self._config['PROXY']:
-                os.remove(_file)
+                os.remove(nk_file)
 
         return _rtcode
 
@@ -331,9 +329,8 @@ class Files(list):
         """Update self from renderable files in dir."""
 
         del self[:]
-        _files = [i for i in os.listdir(
-            os.getcwd()) if i.endswith(('.nk', '.nk.lock'))]
-        _files.sort(key=os.path.getmtime, reverse=False)
+        _files = sorted([get_unicode(i) for i in os.listdir(
+            os.getcwd()) if i.endswith(('.nk', '.nk.lock'))], key=os.path.getmtime, reverse=False)
         self.extend(_files)
 
     def unlock_all(self):
@@ -368,17 +365,48 @@ class Files(list):
         return locked_file
 
     @staticmethod
-    def archive(f, dest='old_files'):
+    def archive(f, dest=u'文件备份'):
         """Archive file in a folder with time struture."""
+
         now = datetime.datetime.now()
         weekday = ('周日', '周一', '周而', '周三', '周四', '周五', '周六')
         dest = os.path.join(
             dest,
-            now.strftime(u'%Y_%m'),
-            unicode(now.strftime(u'%d日_{}#%H时%M分/'))
-        )
-        dest = dest.format(weekday[int(now.strftime('%w'))])
+            get_unicode(now.strftime(u'%Y年%m月')),
+            get_unicode(now.strftime(u'%d日%H时%M分_{}/'))
+        ).format(weekday[int(now.strftime('%w'))])
         copy(f, dest)
+
+    def remove_old_version(self):
+        """Remove all old version nk files.  """
+
+        all_version = {}
+        while True:
+            for i in self:
+                if not os.path.exists(i):
+                    continue
+                shot, version = self.split_version(i)
+                prev_version = all_version.get(shot, -2)
+                if version > prev_version:
+                    all_version[shot] = version
+                    break
+                elif version < prev_version:
+                    self.archive(i)
+                    os.remove(i)
+            else:
+                break
+
+    @staticmethod
+    def split_version(f):
+        """Return nuke style _v# (shot, version number) pair.  """
+
+        match = re.match(r'(.+)_v(\d+)', f)
+        if not match:
+            return (f, -1)
+        shot, version = match.groups()
+        if version < 0:
+            raise ValueError('Negative version number not supported.')
+        return (shot, version)
 
 
 class MainWindow(QMainWindow, Ui_MainWindow):
@@ -391,6 +419,8 @@ class MainWindow(QMainWindow, Ui_MainWindow):
             self.actionNuke.triggered.connect(self.ask_nuke)
             self.actionStop.triggered.connect(self.stop)
             self.actionOpenDir.triggered.connect(self.open_dir)
+            self.actionRemoveOldVersion.triggered.connect(
+                self.remove_old_version)
 
         def _edits():
             for edit, key in self.edits_key.iteritems():
@@ -426,6 +456,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
         self._config = Config()
         self._proc = None
+        self.rendering = False
 
         self.edits_key = {
             self.dirEdit: 'DIR',
@@ -445,7 +476,9 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         Files().unlock_all()
 
     def open_dir(self):
-        url_open('file://{}'.format(self._config['DIR']))
+        """Open dir in explorer."""
+
+        url_open(u'file://{}'.format(self._config['DIR']))
 
     def _start_update(self):
         """Start a thread for update."""
@@ -456,16 +489,19 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
     def update(self):
         """Update UI content."""
+        rendering = bool(self._proc and self._proc.is_alive())
+        if not rendering and self.rendering:
+            QApplication.alert(self)
+            self.statusbar.showMessage(u'渲染已完成')
+        self.rendering = rendering
 
         def _button_enabled():
-            if self._proc and self._proc.is_alive():
-                # Rendering.
+            if rendering:
                 self.renderButton.setEnabled(False)
                 self.stopButton.setEnabled(True)
                 self.listWidget.setStyleSheet(
                     'color:white;background-color:rgb(12%, 16%, 18%);')
             else:
-                # Not rendering.
                 if os.path.isdir(self._config['DIR']) and Files():
                     self.renderButton.setEnabled(True)
                 else:
@@ -474,7 +510,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
                 self.listWidget.setStyleSheet('')
 
         def _edits():
-            for qt_edit, k in self.edits_key.iteritems():
+            for qt_edit, k in self.edits_key.items():
                 try:
                     if isinstance(qt_edit, QtGui.QLineEdit):
                         qt_edit.setText(self._config[k])
@@ -500,8 +536,12 @@ class MainWindow(QMainWindow, Ui_MainWindow):
         dir_ = dialog.getExistingDirectory(
             dir=os.path.dirname(self._config['DIR']))
         if dir_:
-            self._config['DIR'] = dir_
-            self.update()
+            try:
+                dir_.encode('ascii')
+            except UnicodeEncodeError:
+                self.statusbar.showMessage(u'Nuke只支持英文路径', 10000)
+            else:
+                self._config['DIR'] = dir_
 
     def ask_nuke(self):
         """Show a dialog ask config['NUKE']"""
@@ -511,6 +551,7 @@ class MainWindow(QMainWindow, Ui_MainWindow):
             dir=os.getenv('ProgramFiles'), filter='*.exe')[0]
         if filenames:
             self._config['NUKE'] = filenames
+            print('test')
             self.update()
 
     def render(self):
@@ -518,6 +559,10 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
         self._proc = BatchRender()
         self._proc.start()
+
+    def remove_old_version(self):
+        """Remove old version nk files from UI.  """
+        Files().remove_old_version()
 
     def stop(self):
         """Stop rendering from UI."""
@@ -549,7 +594,6 @@ class MainWindow(QMainWindow, Ui_MainWindow):
 
 def main():
     """Run this script standalone."""
-    __file__ = sys.argv[0]
     import fix_pyinstaller
     fix_pyinstaller.main()
     call(u'@TITLE batchrender.console', shell=True)
@@ -587,11 +631,35 @@ def copy(src, dst):
 
 
 def url_open(url):
-    _cmd = "rundll32.exe url.dll,FileProtocolHandler {}".format(url)
-    Popen(_cmd)
+    """Open url in explorer. """
+
+    _cmd = u'rundll32.exe url.dll,FileProtocolHandler "{}"'.format(url)
+    unicode_popen(_cmd)
+
+
+def get_unicode(string, codecs=('UTF-8', OS_ENCODING)):
+    """Return unicode by try decode @string with @codecs.  """
+
+    if isinstance(string, unicode):
+        return string
+
+    for i in codecs:
+        try:
+            return unicode(string, i)
+        except UnicodeDecodeError:
+            continue
+
+
+def unicode_popen(args, **kwargs):
+    """Return Popen object use encoded args.  """
+
+    if isinstance(args, unicode):
+        args = args.encode(OS_ENCODING)
+    return Popen(args, **kwargs)
 
 
 if __name__ == '__main__':
+    __file__ = sys.argv[0]
     try:
         main()
     except SystemExit as ex:
